Add files via upload
I didn't know we need to connect to Github, so I worked on repl. After I finished the web design today, my friend told me that we should link to repl as recod. Therefore, it may not see the detailed change records on GitHub. These files are some drafts before I make the final version and I hope they can be used as the record.